requirements.txt

torch>=2.0.0
torchvision>=0.15.0
torchaudio>=0.10.0
librosa>=0.10.0
soundfile>=0.12.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
tensorboard>=2.13.0